feat: switch to custom modelloader
