--- src/main/java/com/klikli_dev/modonomicon/client/BookBakedModel.java
@@ -1,112 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 klikli-dev
- *
- * SPDX-License-Identifier: MIT
- */
-
-package com.klikli_dev.modonomicon.client;
-
-import com.klikli_dev.modonomicon.book.Book;
-import com.klikli_dev.modonomicon.item.ModonomiconItem;
-import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.multiplayer.ClientLevel;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.block.model.BakedQuad;
-import net.minecraft.client.renderer.block.model.BlockModel;
-import net.minecraft.client.renderer.block.model.ItemOverrides;
-import net.minecraft.client.renderer.block.model.ItemTransforms;
-import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.client.resources.model.BakedModel;
-import net.minecraft.client.resources.model.ModelBaker;
-import net.minecraft.client.resources.model.ModelBakery;
-import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.core.Direction;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.RandomSource;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.client.model.data.ModelData;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.Collections;
-import java.util.List;
-
-public class BookBakedModel implements BakedModel {
-
-    private final BakedModel original;
-    private final ItemOverrides overrides;
-
-    public BookBakedModel(BakedModel original, ModelBakery loader) {
-        this.original = original;
-        BlockModel missing = (BlockModel) loader.getModel(ModelBakery.MISSING_MODEL_LOCATION);
-
-        //TODO: test
-        this.overrides = new ItemOverrides(null, missing, Collections.emptyList()) {
-            @Override
-            public BakedModel resolve(@NotNull BakedModel original, @NotNull ItemStack stack,
-                                      @Nullable ClientLevel world, @Nullable LivingEntity entity, int seed) {
-                Book book = ModonomiconItem.getBook(stack);
-                if (book != null) {
-                    ModelResourceLocation modelPath = new ModelResourceLocation(book.getModel(), "inventory");
-                    return Minecraft.getInstance().getModelManager().getModel(modelPath);
-                }
-                return original;
-            }
-        };
-    }
-
-
-    @Override
-    public List<BakedQuad> getQuads(@Nullable BlockState state, @Nullable Direction side, RandomSource random) {
-        return this.original.getQuads(state, side, random);
-    }
-
-    @Override
-    public @NotNull List<BakedQuad> getQuads(@Nullable BlockState state, @Nullable Direction side, @NotNull RandomSource rand, @NotNull ModelData data, @Nullable RenderType renderType) {
-        return this.original.getQuads(state, side, rand, data, renderType);
-    }
-
-    @Override
-    public boolean useAmbientOcclusion() {
-        return this.original.useAmbientOcclusion();
-    }
-
-    @Override
-    public boolean isGui3d() {
-        return this.original.isGui3d();
-    }
-
-    @Override
-    public boolean usesBlockLight() {
-        return this.original.usesBlockLight();
-    }
-
-    @Override
-    public boolean isCustomRenderer() {
-        return this.original.isCustomRenderer();
-    }
-
-    @NotNull
-    @Override
-    public TextureAtlasSprite getParticleIcon() {
-        return this.original.getParticleIcon();
-    }
-
-    @Override
-    public ItemTransforms getTransforms() {
-        return this.original.getTransforms();
-    }
-
-    @Override
-    public BakedModel applyTransform(ItemTransforms.TransformType transformType, PoseStack poseStack, boolean applyLeftHandTransform) {
-        return this.original.applyTransform(transformType, poseStack, applyLeftHandTransform);
-    }
-
-    @Override
-    public ItemOverrides getOverrides() {
-        return this.overrides;
-    }
-}
